+++ frontend/src/components/ui/editor/editor-ui/color-picker.tsx
@@ -821,11 +821,10 @@ const ColorPickerRoot = React.memo((props: ColorPickerRootProps) => {
 });
 ColorPickerRoot.displayName = "ColorPickerRoot";
 
-interface ColorPickerRootImplProps
-  extends Omit<
-    ColorPickerRootProps,
-    "defaultValue" | "onValueChange" | "format" | "defaultFormat" | "onFormatChange"
-  > {}
+type ColorPickerRootImplProps = Omit<
+  ColorPickerRootProps,
+  "defaultValue" | "onValueChange" | "format" | "defaultFormat" | "onFormatChange"
+>;
 
 function ColorPickerRootImpl(props: ColorPickerRootImplProps) {
   const {
@@ -941,7 +940,7 @@ function ColorPickerRootImpl(props: ColorPickerRootImplProps) {
   );
 }
 
-interface ColorPickerTriggerProps extends React.ComponentProps<typeof PopoverTrigger> {}
+type ColorPickerTriggerProps = React.ComponentProps<typeof PopoverTrigger>;
 
 function ColorPickerTrigger(props: ColorPickerTriggerProps) {
   const { asChild, ...triggerProps } = props;
@@ -956,7 +955,7 @@ function ColorPickerTrigger(props: ColorPickerTriggerProps) {
   );
 }
 
-interface ColorPickerContentProps extends React.ComponentProps<typeof PopoverContent> {}
+type ColorPickerContentProps = React.ComponentProps<typeof PopoverContent>;
 
 function ColorPickerContent(props: ColorPickerContentProps) {
   const { asChild, className, children, ...popoverContentProps } = props;
@@ -1099,7 +1098,7 @@ function ColorPickerArea(props: ColorPickerAreaProps) {
   );
 }
 
-interface ColorPickerHueSliderProps extends React.ComponentProps<typeof SliderPrimitive.Root> {}
+type ColorPickerHueSliderProps = React.ComponentProps<typeof SliderPrimitive.Root>;
 
 function ColorPickerHueSlider(props: ColorPickerHueSliderProps) {
   const { className, ...sliderProps } = props;
@@ -1141,7 +1140,7 @@ function ColorPickerHueSlider(props: ColorPickerHueSliderProps) {
   );
 }
 
-interface ColorPickerAlphaSliderProps extends React.ComponentProps<typeof SliderPrimitive.Root> {}
+type ColorPickerAlphaSliderProps = React.ComponentProps<typeof SliderPrimitive.Root>;
 
 function ColorPickerAlphaSlider(props: ColorPickerAlphaSliderProps) {
   const { className, ...sliderProps } = props;
@@ -1252,7 +1251,7 @@ function ColorPickerSwatch(props: ColorPickerSwatchProps) {
   );
 }
 
-interface ColorPickerEyeDropperProps extends React.ComponentProps<typeof Button> {}
+type ColorPickerEyeDropperProps = React.ComponentProps<typeof Button>;
 
 function ColorPickerEyeDropper(props: ColorPickerEyeDropperProps) {
   const { children, size, ...buttonProps } = props;